# backend/src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BlogSphere API",
    "description": "API Documentation for BlogSphere application",
    "version": "1.0.0"
  },
  "host": "blog-sphere-backend-ruby.vercel.app",
  "schemes": ["https"],
  "basePath": "/",
  "consumes": ["application/json", "multipart/form-data"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter token in format: Bearer <token>"
    }
  },
  "definitions": {
    "ApiResponse": {
      "type": "object",
      "properties": {
        "status": { "type": "integer", "example": 200 },
        "data": { "type": "object" },
        "message": { "type": "string", "example": "Operation successful" }
      }
    },
...
  },
  "paths": {
...
    "/api/v1/users/{userName}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get a single user and their posts",
        "description": "Retrieves a user’s profile and posts by their username.",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "userName",
            "description": "Username of the user",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "User and posts fetched",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/definitions/ApiResponse" }
              }
            }
          },
          "404": { "description": "User not found" }
        }
      }
    },
...
  }
}